Real specs2-based tests.
