+++ src/main/scala/up/KList.scala
@@ -44,16 +44,3 @@ object KList {
   // nicer alias for pattern matching
   val :^: = KCons
 }
-
-object KTest {
-  val f = new (Option ~> List) { def apply[T](o: Option[T]): List[T] = o.toList }
-
-  val x = Some(3) :^: Some("asdf") :^: KNil
-  val y = x map f
-  y match { case List(3) :^: List("asdf") :^: KNil => println("true") }
-
-  val head = new (List ~> Id) { def apply[T](xs: List[T]): T = xs.head }
-  val z = y down head
-  // pattern matching with :^: without 'down' doesn't work (Id again)
-  z match { case 3 :+: "asdf" :+: HNil => println("true") }
-}
