--- src/main/scala/up/Test.scala
@@ -1,13 +1,0 @@
-package up
-
-
-object Test {
-  def main(args: Array[String]) {
-    HListTest.run()
-    HApplyTest.run()
-    NatTest.run()
-    DenseTest.run()
-    //TreeTest.run()
-    TypedMapTest.run()
-  }
-}
