+++ src/test/scala/up/UpdateTest.scala
@@ -1,10 +1,14 @@
 package up
 
-
+import org.specs2.Specification
 import HList._
 import scala.xml.NodeSeq
 
-object UpdateTest {
+class UpdateTest extends Specification {
+  def is = "Update should"   ^
+    "work as expected" ! run ^
+                         end
+
   val h = 3 :: false :: "asdf" :: 'c' :: 9.0 :: List(1,2,3) ::
     Map(1 -> 2, 3 -> 4) :: (<x><asdf/></x> : NodeSeq) :: HNil
 
@@ -19,4 +23,6 @@ object UpdateTest {
   val u1 = h update g1
 
   val u2 = h update g2
+
+  def run = ok
 }
